requirements.txt
jupyter
notebook
sqlalchemy
psycopg2-binary
pandas
pycoingecko
vaderSentiment
praw
python-dotenv
tenacity
